Replace AI::Response with Hash (#13)
Previously, the AI::Response class was nothing more than an initialize message that set up attributes. AI::Response objects in the messages Array are less beginner friendly than Hashes and more difficult to read with amazing print.

This commit removes the AI::Response class and replaces it with a plain Ruby Hash.