Enable background mode (#24)
This commit adds support for API calls to be made in the background.

A large motivation for this feature is to more easily support deep research models or models that can take a long time to respond.

A message (element from the internal messages Array) now has a status attribute, to represent if the request has completed.

Background requests can be manually polled by method call, which updates the message in the Array messages.
Background requests can "auto-poll" by passing an argument. This renders a spinner and will poll automatically at increasing intervals (with a cap at ~110 seconds) until the response status is completed or the duration exceeds the timeout (default 10 minutes).