Add schema generator (#34)
OpenAI no longer shows their Structured Output schema generator in the docs; it's only visible in the Playground, which means you have to sign up and buy credits to use it.

To decrease the amount of friction using Structured Output, we need a way to easily generate JSON schemas with AI::Chat.

Additionally, It's Important to be able to display the generated JSON schema as a nicely indented String for these reasons:
- to show what is being generated
- to allow for easy copy/pasting
- to be consistent with API specifications (String not a Hash)

This commit adds the ability to generate a JSON Schema based on a description with the class method `AI::Chat.generate_schema!`.